+++ CHANGES.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 2.4.3 (2016-05-05)
 - [IMPROVED] Reduced the length of the User-Agent header string.
 - [IMPROVED] Use a more efficient HEAD request for getting revision information when using
   `DesignDocumentManager.remove(String id)`.